UI Update
All UI updated to run working smoothly with each scenario.

Game Over screen added which shows what number of food types your placed in the tower and a burger bun spawner. A button for a tower tour has been added but the functionality has not been setup yet.

Slow motion button added to main menu.

No visual updates added as of yet.

A creative decision has been made to make the tower a burger and not a chopping board. This theme change will be the new direction for the rest of the visuals including later UI updates.