API to level 36 and Audio
Updated API to level 36
Added a calculation to make the game fairer. This was very much needed.
Changed Adaptive app icon to try stop the cropping issue on devices... It did not work and I am yet to find a reason why. This will be looked into in another update.
Added Audio:
1) Cloud popping on loading screen
2) Calm Music
3) Chopping board chopping on slice
4) Increase in score upon on gaining a point without a chop
5) Burger bun squash sound on spawn
6) Burger bun death whistle as it falls to it's death